exceptions: Move InvitationError to zerver/lib/exceptions.
